refactor(renderer): extract shared GatewayFilterDropdown component
The gateway filter UI was duplicated between AgentWorkspaceList and
SecretVaultList. Extract it into lib/gateways/ to reduce duplication
and keep business-logic components separate from upstream-shared lib/ui/.
