+++ packages/renderer/src/lib/secret-vault/SecretVaultList.svelte
@@ -11,7 +11,7 @@ import {
   TableRow,
 } from '@podman-desktop/ui-svelte';
 
-import GatewayFilterDropdown from '/@/lib/ui/GatewayFilterDropdown.svelte';
+import GatewayFilterDropdown from '/@/lib/gateways/GatewayFilterDropdown.svelte';
 import NoLogIcon from '/@/lib/ui/NoLogIcon.svelte';
 import { handleNavigation } from '/@/navigation';
 import {
